% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time.R
\name{add_time}
\alias{add_time}
\title{Add a specific amount of time to a given date with time}
\usage{
add_time(datetime, amount, unit)
}
\arguments{
\item{datetime}{The starting datetime in POSIXct format, YYYY-MM-DD HH:MM:SS.}

\item{amount}{The amount of time to add.}

\item{unit}{The unit of time to add (seconds, minutes, hours, days, weeks).}
}
\value{
The datetime after adding the specific amount of time.
}
\description{
Add a specific amount of time to a given date with time
}
\examples{
add_time("2024-05-21 08:00:00", 1, "hours")  # Add 1 hour
}
