% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_difference.R
\name{date_difference}
\alias{date_difference}
\title{Calculate the time difference between two dates}
\usage{
date_difference(first_date, second_date, units = "days")
}
\arguments{
\item{first_date}{First date in YYYY-MM-DD format.}

\item{second_date}{Second date in YYYY-MM-DD format.}

\item{units}{The unit of time difference. Options are "seconds", "minutes", "hours", "days", "weeks", "months", or "years".}
}
\value{
Time difference between two dates in the filled in unit.
}
\description{
Calculate the time difference between two dates
}
\examples{
date_difference("2024-05-21", "2024-05-22", units = "days")
}
