% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_dutch_holiday.R
\name{is_dutch_holiday}
\alias{is_dutch_holiday}
\title{Check if a date is a holiday in the Netherlands}
\usage{
is_dutch_holiday(date)
}
\arguments{
\item{date}{The date to check in YYYY-MM-DD format.}
}
\value{
TRUE if the date is a holiday, FALSE otherwise.
}
\description{
Check if a date is a holiday in the Netherlands
}
\examples{
is_dutch_holiday("2024-01-01")
is_dutch_holiday("2024-04-27")

}
