% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_same_day.R
\name{is_same_day}
\alias{is_same_day}
\title{Check if two dates fall on the same day}
\usage{
is_same_day(date1, date2)
}
\arguments{
\item{date1}{The first date in YYYY-MM-DD format.}

\item{date2}{The second date in YYYY-MM-DD format.}
}
\value{
A list with the logical value telling whether the dates fall on the same day, and if not, the days of the week on which they fall.
}
\description{
Check if two dates fall on the same day
}
\examples{
is_same_day("2024-05-21", "2024-05-21")
is_same_day("2024-05-04", "2024-02-02")
}
